nerava-backend-v9/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
sqlalchemy==2.0.23
sqlalchemy-utils==0.41.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.6
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0
alembic==1.12.1
stripe>=6.0.0
httpx>=0.24.0
email-validator>=2.0.0
redis>=5.0.0
requests>=2.31.0
prometheus-client>=0.19.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
qrcode[pil]>=7.4.2
cryptography>=41.0.0
google-auth>=2.23.0
phonenumbers>=8.13.0
twilio>=8.10.0
